+++ Modules/Intent.Modules.Common.CSharp/Nuget/NugetRegistry.cs
@@ -25,6 +25,10 @@ namespace Intent.Modules.Common.CSharp.Nuget
 
         public static NugetPackageInfo GetVersion(string packageName, MajorMinorVersion framework)
         {
+            if (packageName == "MediatR")
+            {
+
+            }
             if (!_registrations.TryGetValue(packageName, out var registration))
             {
                 throw new Exception($"NuGet package not registered : `{packageName}`");
@@ -42,14 +46,39 @@ namespace Intent.Modules.Common.CSharp.Nuget
             var lockedVersions = matchedVersions.Where(v => v.Locked).ToList();
             if (lockedVersions.Any())
             {
-                var lowestLocked = lockedVersions.OrderBy(v => v.Version).First();
-                return new NugetPackageInfo(packageName, lowestLocked.Version.ToString());
+                var lowestLockedVersion = lockedVersions.Min(v => v.Version);
+
+                return GetNugetPackageInfo(packageName, lowestLockedVersion, lockedVersions.Where(x => x.Version == lowestLockedVersion));
             }
             else
             {
-                var highestUnlocked = matchedVersions.OrderByDescending(v => v.Version).First();
-                return new NugetPackageInfo(packageName, highestUnlocked.Version.ToString());
+                var highestVersion = matchedVersions.Max(v => v.Version);
+                return GetNugetPackageInfo(packageName, highestVersion, matchedVersions.Where(x => x.Version == highestVersion));
             }
+        }
+
+        private static NugetPackageInfo GetNugetPackageInfo(string packageName, NuGetVersion version, IEnumerable<PackageVersion> versionInfos)
+        {
+            var result =  new NugetPackageInfo(packageName, version.ToString());
+            var includeAssets = versionInfos.SelectMany(x => x.IncludeAssets).Distinct();
+            var privateAssets = versionInfos.SelectMany(x => x.PrivateAssets).Distinct();
+            if (includeAssets.Any() || privateAssets.Any())
+            {
+                result.SpecifyAssetsBehaviour(privateAssets, includeAssets);
+            }
+
+            var consolidatedDependencies = versionInfos.SelectMany(x => x.Dependencies)
+                .GroupBy(d => d.Name)
+                .Select(g => g.First()) // Version should be identical as they are dependents for the same NuGet Package
+                .ToList();
+            if (consolidatedDependencies.Any())
+            {
+                foreach (var dependency in consolidatedDependencies)
+                {
+                    result.WithNugetDependency(dependency.Name, dependency.Version);
+                }
+            }
+            return result;
         }
 
         private class NugetPackageRegistrations
@@ -62,18 +91,6 @@ namespace Intent.Modules.Common.CSharp.Nuget
 
             public string PackageName { get; }
             public List<Func<MajorMinorVersion, PackageVersion>> RegisteredVersions { get; }
-        }
-    }
-
-    public class PackageVersion
-    {
-
-        public bool Locked { get; }
-        public NuGetVersion Version { get; }
-        public PackageVersion(string version, bool locked = false)
-        {
-            Version = new NuGetVersion(version);
-            Locked = locked;
         }
     }
 }