+++ Modules/Intent.Modules.ModuleBuilder.CSharp/Tasks/NuGetApi.cs
@@ -68,7 +68,8 @@ namespace Intent.Modules.ModuleBuilder.CSharp.Tasks
 
                 if (latestPackage != null)
                 {
-                    result.Add(new NugetVersionInfo(framework, latestPackage.Identity.Version));
+                    var dependencies = latestPackage.DependencySets.FirstOrDefault(d => d.TargetFramework == framework);
+                    result.Add(new NugetVersionInfo(framework, latestPackage.Identity.Version, dependencies.Packages.Select(p => new NugetDependencyInfo(p.Id, p.VersionRange)).ToList()));
                 }
             }
             return result;
@@ -76,14 +77,30 @@ namespace Intent.Modules.ModuleBuilder.CSharp.Tasks
 
         internal class NugetVersionInfo
         {
-            public NugetVersionInfo(NuGetFramework frameworkVersion, NuGetVersion packageVersion)
+            public NugetVersionInfo(NuGetFramework frameworkVersion, NuGetVersion packageVersion, List<NugetDependencyInfo> dependencies)
             {
                 FrameworkVersion = frameworkVersion;
                 PackageVersion = packageVersion;
+                Dependencies = dependencies;
             }
 
             public NuGetFramework FrameworkVersion { get; }
             public NuGetVersion PackageVersion { get; }
+
+            public List<NugetDependencyInfo> Dependencies { get; }
+
+        }
+
+        internal class NugetDependencyInfo
+        {
+            public NugetDependencyInfo(string packageName, VersionRange version)
+            {
+                PackageName = packageName;
+                Version = version;
+            }
+
+            public string PackageName { get; }
+            public VersionRange Version { get; } 
 
         }
     }